@@ -3,14 +3,13 @@ apply plugin: 'com.github.dcendents.android-maven'
 group='com.github.liyuzhao'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.3"
+    compileSdkVersion 28
 
     defaultConfig {
         minSdkVersion 14
-        targetSdkVersion 27
-        versionCode 1
-        versionName "1.0"
+        targetSdkVersion 28
+        versionCode 20
+        versionName "2.0"
     }
     buildTypes {
         release {
