@@ -1,7 +1,9 @@
 package com.bigkoo.pickerview;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
+
+import com.easemob.pickerview.R;
 
 public class MainActivity extends AppCompatActivity {
 
